==> phenotypes/scripts/S01_ColdTolerance/S01_CalculateNormalisedSurvivalAverages.Rmd <==
---
title: "Calculate Normalised Survival Fractions"
author: "FiG-T"
date: "`r Sys.Date()`"
header-includes:
   - \usepackage[default]{sourcesanspro}
   - \usepackage[T1]{fontenc}
mainfont: SourceSansPro
output: 
  html_document: 
    keep_md: yes
    toc: yes
toc: TRUE
editor_options:
  markdown:
    wrap: 80
  chunk_output_type: inline
---

## Introduction 

One measure of fitness used in this experiment is survival (following stress treatment). This is designed to approximate how well a fly can survive without the need for a full longevity assay (which may take months), and is thus more suitable for a study of this size.

Flies from different populations (1-11,14) are left in a chill coma state for 25h (~4ºC) - this is used as the stress treatment. Upon removal the flies are returned to room temperature and left to recover. The number of surviving flies per vial (of an original 10) are then counted after 5h (survival_5h) and 24h (survival_24h).

> *Libraries* : Libraries are now listed in the S01_PathsShortcuts.R config file.  This is to make it easier to see which libraries are used across the whole project. 

## Setup

This script calls information in from the project-specific setup these files. These contain information on the relative paths to different folders and the colour palettes.

```{r configure-setup, echo=FALSE, message=FALSE, warning=FALSE }
# source all files in configure directory
for (file in list.files(here::here("phenotypes/scripts/S00_ConfigureSetup"))) {
  source(here::here("phenotypes/scripts/S00_ConfigureSetup", file), verbose = FALSE)
}
rm(file)

# set path to data 
path_to_data <- here::here("phenotypes/data/D01_ColdTolerance/")

# set path to results for this file
path_to_results <- here::here("phenotypes/results/R01_ColdTolerance/")
```

## Load Data

```{r load-data}
# load data 
cold_survival <- readxl::read_excel(
  path = here(path_to_data, "TIME_thermal_survival_data.xlsx"), 
  trim_ws = TRUE, 
  col_names = TRUE, 
  range = readxl::cell_cols("A:J")
)
```

### Format Factor Levels

```{r format-factor-levels}
# set the order of genotypes
cold_survival$genotype <- factor(
  cold_survival$genotype, 
  levels = genotype_factor_levels
)

# set the order of populations
cold_survival$pop <- factor(
  cold_survival$pop, 
  levels = population_factor_levels
)
```

If this test fails then it is likely some vials have more flies at the 5h timepoint than they did at the start (which should never happen). This is likely a data-input error and will need to be treated. 

```{r check-survival-limits}
# run test
testthat::expect_equal(
  nrow(cold_survival |> dplyr::filter(survival_5h > start)), 
  0 # the tibble should be empty
)
```

## Calculate Survival Fractions

```{r calculate-survival-fractions}
cold_survival <- cold_survival |> 
  dplyr::mutate(
    # percentage survival 5 hours after removal from treatment
    per_5h = (survival_5h / start) * 100,
    # percentage survival 24 hours after removal from treatment
    per_24h = (survival_24h /start) * 100, 
    # survival percentage only including those that survived cold treatment
    per_drop = (survival_24h/survival_5h) * 100
  )
```

## Calculate Control Averages Per Timepoint

```{r population-averages}
cold_survival <- cold_survival |> 
  filter(genotype %in% c("mM", "tT")) |> 
  group_by(timepoint) |>
  reframe(control_ave = mean(per_drop)) |>
  left_join(
    x = cold_survival, 
    y = _ , 
    by = "timepoint"
  )
```

## Normalise *Each Vial* By Control Average for the Timepoint

```{r grouped-averages, warning=FALSE}
cold_survival_normalised <- cold_survival |> 
  group_by(genotype, pop, timepoint, replicate) |>
  reframe(
    pop = as.factor(pop),
    initial_temp = as.numeric(initial_temp), 
    date = date,
    tray = tray,
    control_ave = control_ave,
    survival_drop = per_drop,
    survival_drop_normalised = per_drop / control_ave
  ) |> 
  dplyr::distinct()
```

```{r save-data}
readr::write_csv(
  x = cold_survival_normalised, 
  file = here(path_to_data, "normalised_survival.csv"), 
  col_names = TRUE
)
```

## Calculate Grouped Averages

```{r grouped-averages, warning=FALSE}
cold_survival_normalised_averages <- cold_survival_normalised |> 
  group_by(genotype, pop, timepoint) |>
  reframe(
    pop = as.factor(pop),
    initial_temp = initial_temp, 
    date = date,
    control_ave = control_ave,
    survical_drop_sd = sd(survival_drop, na.rm = TRUE),
    survical_drop = mean(survival_drop, na.rm = TRUE),
    survival_drop_normalised_sd = sd(survival_drop_normalised, na.rm = TRUE),
    survival_drop_normalised = mean(survival_drop_normalised, na.rm = TRUE), 
  ) |> 
  dplyr::distinct()
```

## Save Formatted Data 

```{r save-data}
readr::write_csv(
  x = cold_survival_normalised_averages, 
  file = here(path_to_data, "normalised_survial_population_averages.csv"), 
  col_names = TRUE
)
```
