==> phenotypes/scripts/S01_ColdTolerance/S04_StatisticsNormalisedSurvivalFractions.Rmd <==
---
title: " Statistics: Normalised Survival Fractions"
author: "FiG-T"
date: "`r Sys.Date()`"
header-includes:
   - \usepackage[default]{sourcesanspro}
   - \usepackage[T1]{fontenc}
mainfont: SourceSansPro
output: 
  html_document: 
    keep_md: yes
    toc: yes
toc: TRUE
editor_options:
  markdown:
    wrap: 80
  chunk_output_type: inline
---

## Introduction 

Statistics for trends in cold survival (across timepoints and between genotypes). The input file for this script is created at the end of 'S01_CalculateNormalisedSurvivalAverages.Rmd'. 

> *Libraries* : Libraries are now listed in the S01_PathsShortcuts.R config file.  This is to make it easier to see which libraries are used across the whole project. 

## Setup

This chunk calls information in from the project-specific setup these files. These contain information on the relative paths to different folders and the colour palettes.

```{r configure-setup, echo=FALSE, message=FALSE, warning=FALSE }
# source all files in configure directory
for (file in list.files(here::here("phenotypes/scripts/S00_ConfigureSetup"))) {
  source(here::here("phenotypes/scripts/S00_ConfigureSetup", file), verbose = FALSE)
}
rm(file)

# set path to data 
path_to_data <- here::here("phenotypes/data/D01_ColdTolerance/")

# set path to results for this file
path_to_results <- here::here("phenotypes/results/R01_ColdTolerance/R04_StatisticsNormalisedSurvivalFractions")
```

## Load Data

The first approach is using the normalised, but un-grouped data. 

```{r load-data}
# load data 
cold_survival_normalised <- readr::read_csv(
  file = here(path_to_data, "normalised_survival.csv"), 
  col_names = TRUE, 
  show_col_types = FALSE
)
```

## Scale variables 

As some variables have a much wider range than others - the plan is to scale the initial temperature variable

```{r scale-variables}

cold_survival_normalised <- cold_survival_normalised |> 
  dplyr::mutate(
    timepoint = as.factor(timepoint), 
    genotype = as.factor(genotype), 
    pop = as.factor(pop), 
    initial_temp_scaled = scale(initial_temp), 
    date = as.factor(date), 
    tray = as.factor(tray)
  )
  
```

## Build Models

```{r build-models}
# build mixed effects models: 
survival_model <-  lmerTest::lmer(
  data = cold_survival_normalised , #|> filter(genotype %in% c("mX", "tX")),
  formula = survival_drop_normalised ~  pop * timepoint  + (1|tray)
)

# run anova with type 3 correction to account for potentially significant interaction terms. 
anova(survival_model, type = 3)

```

To check that assumptions are not massively violated - check the distribution of model residuals:

```{r plot-residuals}
ggplot(
  mapping = aes(
    sample = residuals(survival_model)
  )
) +
  stat_qq_line(
  ) +
  stat_qq() +
  theme_light()
  
```
The above plot shows broad meeting of requirements - I'm always a little unsure as to what is an 'acceptable' level of divergence...  Working on the premise that this is fine, we can then have a look at the estimated marginal means and the pairwise differences between groups. 

```{r calculate-estimated-marginal-means}
emmeans_survival <- emmeans::emmeans(
  survival_model, 
   ~  timepoint * pop 
)
emmeans_survival
```

```{r}
ggplot(
  data = tibble::as_tibble(emmeans_survival), 
  mapping = aes(
    x = timepoint, 
    y = emmean,
    colour = genotype
  )
) +
  geom_point(
    mapping = aes(
      colour = genotype
    ),
    size = 3, 
    show.legend = FALSE
  ) +
  geom_errorbar(
    mapping = aes(
      ymax = emmean + SE, 
      ymin = emmean - SE
    ), 
    width = 0.2
  ) +
  geom_smooth(
     mapping = aes(
      colour = genotype
    ),
    method = "lm", 
    show.legend = FALSE, 
    se = FALSE
  ) +
  scale_colour_manual(
    values = aus_fly_palette
  ) +
  facet_wrap("genotype") +
  labs(
    x = "Timepoint", 
    y = "Estimated Mean (Normalised) Survival"
  ) +
  theme_light() +
  muted_strips


```

```{r calculate pairwise-comparisons}
# using the lmerTest package 
ls_means(survival_model, which = "genotype", pairwise = TRUE)

pairwise_comparisons <- as_tibble(pairs(
  emmeans_survival, 
  simple = "pop"
))

pairwise_comparisons <-  pairwise_comparisons |> 
  tidyr::separate_wider_delim(
    cols = contrast, 
    names = c("popA", "popB"), 
    delim = " - ", 
    
  )

# set the order of populations
pairwise_comparisons$popA <- factor(
  pairwise_comparisons$popA, 
  levels = stringr::str_c("pop", population_factor_levels)
)
pairwise_comparisons$popB <- factor(
  pairwise_comparisons$popB, 
  levels = stringr::str_c("pop", population_factor_levels)
)
```

```{r plot-pairwise-comparisons}
ggplot(
) +
  geom_tile(
    data = pairwise_comparisons |> dplyr::filter(popB %in% c("pop11", "pop14")& p.value > 0.05), 
    mapping = aes(
      x = popA, 
      y = popB, 
      fill = p.value
    )
  ) +
  scale_fill_gradient(
    low = "wheat4", high = "white",
    na.value = "grey90"
    ) +
  geom_tile(
    data = pairwise_comparisons |> dplyr::filter(popB %in% c("pop11", "pop14") & p.value < 0.05), 
    mapping = aes(
      x = popA, 
      y = popB
    ), 
    fill = "orange"
  ) +
  facet_grid(
    cols = vars(timepoint)
  ) +
  theme_light() +
  muted_strips +
  theme(
    axis.title = element_blank(), 
    axis.text.x = element_text(angle = 90)
  )
```

```{r save-normalised-pairwise-comparisons-plot}
ggsave(
  filename = here(path_to_results, "F01_NormalisedSurvivalFractionsVsStartingTemperature.pdf"), 
  plot = last_plot(), 
  bg = "transparent", 
  width = 10, 
  height = 7
)
```
